Push modified xz PKGBUILD that downgrades to 5.5.2beta with epoch.
Arch has declared that their XZ package to be unaffected by CVE-2024-3094,
but out of precautions I downgraded locally-installed XZ utils to 5.5.2beta
with disabled GNU_IFUNC implementation (which is present since 5.5 pre-release)
prior to the actual backdoor being added with 5.6.0 release that makes use of
this feature.

There are even discussions around that we should downgrade to an even older
version of XZ utils where the actor hadn't started contributing upstream yet,
which was 5.2.5 and 5.3.2 pre-release. But for now, until more things surface
about whether there are more backdoors or attempt to insert security holes into
the project this will suffice.